fix test and create helpers
